Refactored startup scripts for robustness - fixed silent start errors due permission issues - fixed infinite loop conditions in shutdown scripts - added option to clean-database.sh to confirm operation - start up scripts now work with configs on different ports - made thrift targets depend on actual generated output instead of fixed list

+++ bin/clean-database.sh
@@ -19,39 +19,53 @@
 export HYPERTABLE_HOME=$(cd `dirname "$0"`/.. && pwd)
 . $HYPERTABLE_HOME/bin/ht-env.sh
 
-# Stop servers other than dfsbroker
-stop_server $HYPERTABLE_HOME/run/ThriftBroker.pid
-stop_server $HYPERTABLE_HOME/run/Hypertable.RangeServer.pid
-stop_server $HYPERTABLE_HOME/run/Hypertable.Master.pid
-stop_server $HYPERTABLE_HOME/run/Hyperspace.pid
-sleep 1
-wait_for_server_shutdown thriftbroker "thrift broker" &
-wait_for_server_shutdown rangeserver "range server" &
-wait_for_server_shutdown master "hypertable master" &
-wait_for_server_shutdown hyperspace "hyperspace" &
+confirm=y
 
-#
-# Clear state
-#
-$HYPERTABLE_HOME/bin/serverup --silent dfsbroker
-if [ $? != 0 ] ; then
-  echo "ERROR: DfsBroker not running, database not cleaned"
-  # remove local stuff anyway.
-  rm -rf $HYPERTABLE_HOME/hyperspace/* $HYPERTABLE_HOME/fs/*
-  exit 1
+if tty > /dev/null && [ $# == 1 ]; then
+  case $1 in
+    -i|--interactive)
+      echo "Are you sure you want to clear the database (on default ports)? (y/N)"
+      read confirm
+      shift
+      ;;
+  esac
 fi
 
-$HYPERTABLE_HOME/bin/dfsclient --eval "rmdir /hypertable/servers"
-$HYPERTABLE_HOME/bin/dfsclient --eval "rmdir /hypertable/tables"
-echo "Removed /hypertable/servers in DFS"
-echo "Removed /hypertable/tables in DFS"
-/bin/rm -rf $HYPERTABLE_HOME/hyperspace/*
-echo "Cleared hyperspace"
+# Stop servers other than dfsbroker
+stop_server thriftbroker rangeserver master hyperspace
+sleep 1
+wait_for_server_shutdown thriftbroker "thrift broker" "$@" &
+wait_for_server_shutdown rangeserver "range server" "$@" &
+wait_for_server_shutdown master "hypertable master" "$@" &
+wait_for_server_shutdown hyperspace "hyperspace" "$@" &
+
+case $confirm in
+  y|Y)
+    #
+    # Clear state
+    #
+    check_server dfsbroker "$@"
+    if [ $? != 0 ] ; then
+      echo "ERROR: DfsBroker not running, database not cleaned"
+      # remove local stuff anyway.
+      rm -rf $HYPERTABLE_HOME/hyperspace/* $HYPERTABLE_HOME/fs/*
+      exit 1
+    fi
+
+    $HYPERTABLE_HOME/bin/dfsclient --eval "rmdir /hypertable/servers" "$@"
+    $HYPERTABLE_HOME/bin/dfsclient --eval "rmdir /hypertable/tables" "$@"
+    echo "Removed /hypertable/servers in DFS"
+    echo "Removed /hypertable/tables in DFS"
+    /bin/rm -rf $HYPERTABLE_HOME/hyperspace/*
+    echo "Cleared hyperspace"
+    ;;
+  *) echo "Database not cleared";;
+esac
 
 #
 # Stop dfsbroker
 #
-stop_server $HYPERTABLE_HOME/run/DfsBroker.*.pid
+stop_server dfsbroker
 sleep 1
-wait_for_server_shutdown dfsbroker "DFS broker" &
+wait_for_server_shutdown dfsbroker "DFS broker" "$@" &
 wait
